Correct function and signature to refresh token
The README.md mentions the existence of a .refresh_token function but none by that name can be found. Therefore I replaced it with refresh_access_token. Also, as the example mentions using the 'saved' refresh token I've added the from_yaml=True argument as this will get the yaml loaded/saved if token changed.